[stdlib/3/multiprocessing] Add queues.pyi, make the other stubs use it
