+++ stdlib/3/multiprocessing/__init__.pyi
@@ -2,7 +2,7 @@
 
 from typing import (
     Any, Callable, ContextManager, Iterable, Mapping, Optional, Dict, List,
-    Union, TypeVar, Sequence, Tuple
+    Union, Sequence, Tuple
 )
 
 from logging import Logger
@@ -12,10 +12,8 @@ from multiprocessing.context import (
     ProcessError, BufferTooShort, TimeoutError, AuthenticationError)
 from multiprocessing.managers import SyncManager
 from multiprocessing.process import current_process as current_process
-import queue
+from multiprocessing.queues import Queue, SimpleQueue, JoinableQueue
 import sys
-
-_T = TypeVar('_T')
 
 # N.B. The functions below are generated at runtime by partially applying
 # multiprocessing.context.BaseContext's methods, so the two signatures should
@@ -61,19 +59,6 @@ class Process():
     def terminate(self) -> None: ...
     def is_alive(self) -> bool: ...
     def join(self, timeout: Optional[float] = ...) -> None: ...
-
-class Queue(queue.Queue[_T]):
-    def __init__(self, maxsize: int = ...) -> None: ...
-    def get(self, block: bool = ..., timeout: Optional[float] = ...) -> _T: ...
-    def put(self, item: _T, block: bool = ..., timeout: Optional[float] = ...) -> None: ...
-    def qsize(self) -> int: ...
-    def empty(self) -> bool: ...
-    def full(self) -> bool: ...
-    def put_nowait(self, item: _T) -> None: ...
-    def get_nowait(self) -> _T: ...
-    def close(self) -> None: ...
-    def join_thread(self) -> None: ...
-    def cancel_join_thread(self) -> None: ...
 
 class Value():
     value: Any = ...
