+++ stdlib/3/multiprocessing/context.pyi
@@ -3,6 +3,7 @@
 from logging import Logger
 import multiprocessing
 from multiprocessing import synchronize
+from multiprocessing import queues
 import sys
 from typing import (
     Any, Callable, Iterable, Optional, List, Mapping, Sequence, Tuple, Type,
@@ -51,12 +52,9 @@ class BaseContext(object):
     def RLock(self) -> synchronize.RLock: ...
     def Semaphore(self, value: int = ...) -> synchronize.Semaphore: ...
 
-    # TODO: change return to Queue once a stub exists in multiprocessing.queues
-    def Queue(self, maxsize: int = ...) -> Any: ...
-    # TODO: change return to Queue once a stub exists in multiprocessing.queues
-    def JoinableQueue(self, maxsize: int = ...) -> Any: ...
-    # TODO: change return to SimpleQueue once a stub exists in multiprocessing.queues
-    def SimpleQueue(self) -> Any: ...
+    def Queue(self, maxsize: int = ...) -> queues.Queue: ...
+    def JoinableQueue(self, maxsize: int = ...) -> queues.JoinableQueue: ...
+    def SimpleQueue(self) -> queues.SimpleQueue: ...
     def Pool(
         self,
         processes: Optional[int] = ...,
